Update to support go version 1.18
We have internal dependency on testing package that was changed in go
version 1.18
